Create index to export components
